+++ VideoSystemWeb/DAL/Dati_Articoli_Lavorazione_DAL.cs
@@ -80,7 +80,7 @@ namespace VideoSystemWeb.DAL
                                         datiArticoliLavorazione.UsaCostoFP = riga.Field<bool?>("usaCostoFP");
                                         datiArticoliLavorazione.Consuntivo = riga.Field<bool?>("consuntivo");
                                         datiArticoliLavorazione.NumOccorrenza = riga.Field<int>("numOccorrenza");
-                                        datiArticoliLavorazione.Quantita = riga.Field<int>("quantita");
+                                        datiArticoliLavorazione.Quantita = riga.Field<decimal>("quantita");
                                         datiArticoliLavorazione.PrezzoUnitario = riga.Field<decimal>("prezzoUnitario");
 
                                         listaDatiArticoli.Add(datiArticoliLavorazione);
